fix windows werror

+++ cocos/3d/skeletal-animation/SkeletalAnimationUtils.cpp
@@ -176,7 +176,7 @@ cc::optional<IJointTextureHandle> JointTexturePool::getDefaultPoseTexture(Skelet
     const std::vector<Mat4> &       bindPoses = skeleton->getBindposes();
     Float32Array                    textureBuffer;
     bool                            buildTexture = false;
-    uint32_t                        jointCount   = joints.size();
+    uint32_t                        jointCount   = static_cast<uint32_t>(joints.size());
     if (!texture.has_value()) {
         uint32_t             bufSize = jointCount * 12;
         ITextureBufferHandle handle;
@@ -227,7 +227,7 @@ cc::optional<IJointTextureHandle> JointTexturePool::getDefaultPoseTexture(Skelet
     }
 
     std::vector<geometry::AABB> bounds;
-    texture->bounds[mesh->getHash()] = bounds;
+    texture->bounds[static_cast<uint32_t>(mesh->getHash())] = bounds;
     geometry::AABB::fromPoints(v3Min, v3Max, &bounds[0]);
     if (buildTexture) {
         _pool->update(texture->handle, textureBuffer.buffer());
